+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/TagAssetFilterDeserializer.java
@@ -1,0 +1,40 @@
+package com.strangegrotto.wealthdraft.assetallocation.datamodel.filters;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
+import com.strangegrotto.wealthdraft.assets.definition.CustomTagDefinition;
+import org.immutables.value.Value;
+
+import java.io.IOException;
+import java.util.Map;
+
+public class TagAssetFilterDeserializer extends JsonDeserializer<TagAssetFilter> {
+    private static class RawTagAssetFilter {
+        @JsonProperty("tag")
+        public String tag;
+
+        @JsonProperty("value")
+        public String value;
+    }
+
+    private final Map<String, CustomTagDefinition> customTags;
+
+    public TagAssetFilterDeserializer(Map<String, CustomTagDefinition> customTags) {
+        this.customTags = customTags;
+    }
+
+    @Override
+    public TagAssetFilter deserialize(JsonParser p, DeserializationContext ctxt) throws IOException, JsonProcessingException {
+        var raw = p.readValueAs(RawTagAssetFilter.class);
+        return ImmTagAssetFilter.of(
+            this.customTags,
+            raw.tag,
+            raw.value
+        );
+    }
+}