+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/TargetAssetAllocationsTest.java
@@ -1,9 +1,11 @@
 package com.strangegrotto.wealthdraft.assetallocation.datamodel;
 
+import com.fasterxml.jackson.databind.exc.ValueInstantiationException;
 import com.strangegrotto.wealthdraft.Main;
 import com.strangegrotto.wealthdraft.assetallocation.ExpectedExampleTargetAssetAllocations;
 import com.strangegrotto.wealthdraft.assetallocation.TargetAssetAllocationsTestFiles;
-import com.strangegrotto.wealthdraft.assetallocation.datamodel.TargetAssetAllocations;
+import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitions;
+import com.strangegrotto.wealthdraft.assets.definition.AssetDefinitionsTestFiles;
 import com.strangegrotto.wealthdraft.assets.definition.ExpectedExampleAssetDefinitions;
 import org.junit.Assert;
 import org.junit.Test;
@@ -11,7 +13,7 @@ import org.junit.Test;
 import java.io.IOException;
 import java.util.Set;
 
-public class TargetAssetAllocationTest {
+public class TargetAssetAllocationsTest {
     @Test
     public void testValidDeserialization() throws IOException {
         var targetAssetAllocations = parseAssetAllocationsFile(TargetAssetAllocationsTestFiles.EXAMPLE);
@@ -33,8 +35,32 @@ public class TargetAssetAllocationTest {
         Assert.assertEquals(expectedAssetIds, filteredAssetIds);
     }
 
+    @Test(expected = ValueInstantiationException.class)
+    public void testErrorOnNonexistentNumeratorFilter() throws IOException {
+        parseAssetAllocationsFile(TargetAssetAllocationsTestFiles.NONEXISTENT_NUMERATOR_FILTER);
+    }
+
+    @Test(expected = ValueInstantiationException.class)
+    public void testErrorOnNonexistentDenominatorFilter() throws IOException {
+        parseAssetAllocationsFile(TargetAssetAllocationsTestFiles.NONEXISTENT_DENOMINATOR_FILTER);
+    }
+
+    @Test(expected = ValueInstantiationException.class)
+    public void testErrorOnFractionGreaterThan1() throws IOException {
+        parseAssetAllocationsFile(TargetAssetAllocationsTestFiles.FRACTION_GREATER_THAN_1);
+    }
+
+    @Test(expected = ValueInstantiationException.class)
+    public void testErrorOnFractionLessThan0() throws IOException {
+        parseAssetAllocationsFile(TargetAssetAllocationsTestFiles.FRACTION_LESS_THAN_0);
+    }
+
     private static TargetAssetAllocations parseAssetAllocationsFile(TargetAssetAllocationsTestFiles testFile) throws IOException {
         var mapper = Main.getObjectMapper();
+        var assetsUrl = AssetDefinitionsTestFiles.EXAMPLE.getResource();
+        var assetDefinitions = mapper.readValue(assetsUrl, AssetDefinitions.class);
+        Main.addDeserializersNeedingAssetDefs(mapper, assetDefinitions);
+
         var assetAllocationsUrl = testFile.getResource();
         return mapper.readValue(assetAllocationsUrl, TargetAssetAllocations.class);
     }