+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/ConjunctionAssetFilterTest.java
@@ -2,6 +2,7 @@ package com.strangegrotto.wealthdraft.assetallocation.datamodel.filters;
 
 import com.strangegrotto.wealthdraft.assetimpls.bankaccount.ImmBankAccountAsset;
 import com.strangegrotto.wealthdraft.assets.definition.Asset;
+import com.strangegrotto.wealthdraft.assets.definition.ImmCustomTagDefinition;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -20,6 +21,15 @@ public class ConjunctionAssetFilterTest {
 
         var matchingAssetId = "matches";
 
+        var unrelatedTagName = "foo";
+        var unrelatedTagValue = "bar";
+
+        var customTags = Map.of(
+                needle1Name, ImmCustomTagDefinition.builder().build(),
+                needle2Name, ImmCustomTagDefinition.builder().build(),
+                unrelatedTagName, ImmCustomTagDefinition.builder().build()
+        );
+
         var haystack = Map.<String, Asset<?, ?>>of(
                 "needle1-but-not-2", ImmBankAccountAsset.of("Needle 1 but not 2").withCustomTags(Map.of(
                         needle1Name, needle1Value
@@ -31,12 +41,14 @@ public class ConjunctionAssetFilterTest {
                 "needle2-but-not-1", ImmBankAccountAsset.of("Needle 2 but not 1").withCustomTags(Map.of(
                         needle2Name, needle2Value
                 )),
-                "unmatching-tags", ImmBankAccountAsset.of("Unmatching tags").withCustomTags(Map.of("foo", "bar")),
+                "unmatching-tags", ImmBankAccountAsset.of("Unmatching tags").withCustomTags(Map.of(
+                        unrelatedTagName, unrelatedTagValue
+                )),
                 "no-tags", ImmBankAccountAsset.of("No tags")
         );
 
-        var needle1Filter = ImmTagAssetFilter.of(needle1Name, needle1Value);
-        var needle2Filter = ImmTagAssetFilter.of(needle2Name, needle2Value);
+        var needle1Filter = ImmTagAssetFilter.of(customTags, needle1Name, needle1Value);
+        var needle2Filter = ImmTagAssetFilter.of(customTags, needle2Name, needle2Value);
         var conjunctiveFilter = ImmConjunctionAssetFilter.of(List.of(
                 needle1Filter,
                 needle2Filter