+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/TargetAssetAllocations.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.base.Preconditions;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
 import com.strangegrotto.wealthdraft.assetallocation.datamodel.filters.AssetFilter;
+import com.strangegrotto.wealthdraft.assets.definition.CustomTagDefinition;
 import org.immutables.value.Value;
 
 import java.math.BigDecimal;
@@ -27,6 +28,7 @@ public interface TargetAssetAllocations {
     default void check() {
         var filters = getFilters();
         var targets = getTargets();
+
         for (int i = 0; i < targets.size(); i++) {
             var target = targets.get(i);