+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.12.1
+* Add extra tests to ensure nonsensical target asset allocation rules are rejected
+* Add tests to make sure that the TagAssetFilter is using only recognized tags and values, to guard against typos
+* Show numerator/denominator value in asset allocation chart
+
 ## 0.12.0
 * Refactor assets history from `assetId -> (date -> assetSnapshot)` to `date -> (assetId -> assetSnapshot)`